llm-server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.0
llama-cpp-python>=0.3
python-dotenv>=1.0
